fix: dropdown 로직, 별도의 hook으로 분리
