+++ src/components/common/Dropdown/Dropdown.tsx
@@ -4,50 +4,24 @@ import { ChevronDown, ChevronUp } from '../../../assets';
 
 import * as S from './Dropdown.style';
 import DropdownOptions from './DropdownOptions';
+import useDropdown from '../../../hooks/useDropdown';
 
 interface DropdownProps<T> {
   optionList: [T, string][];
   onChange: (value: T) => void;
 }
 
-// TODO: dropdown 관련 로직 hook으로 분리
-
 function Dropdown<T extends string>({
   optionList,
   onChange,
 }: DropdownProps<T>) {
-  const [isOpened, setIsOpened] = useState(false);
-  const [preview, setPreview] = useState(optionList[0][1]);
   const dropdownRef = useRef<HTMLDivElement>(null);
-
-  const handleToggleDropdown = () => {
-    setIsOpened(!isOpened);
-  };
-
-  const handleChangeOption = ([key, value]: [T, string]) => {
-    onChange(key);
-    setPreview(value);
-  };
-
-  const handleClickOutside = (event: MouseEvent) => {
-    if (
-      dropdownRef.current &&
-      !dropdownRef.current.contains(event.target as Node)
-    ) {
-      setIsOpened(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener('click', handleClickOutside);
-    return () => {
-      document.removeEventListener('click', handleClickOutside);
-    };
-  }, []);
+  const { isOpened, preview, handleChangeOption, handleToggleDropdown } =
+    useDropdown({ optionList, onChange, dropdownRef });
 
   return (
-    <S.Dropdown ref={dropdownRef}>
-      <S.Preview isOpened={isOpened} onClick={handleToggleDropdown}>
+    <S.Dropdown ref={dropdownRef} onClick={handleToggleDropdown}>
+      <S.Preview isOpened={isOpened}>
         <S.PreviewText>{preview}</S.PreviewText>
         {isOpened ? <ChevronUp /> : <ChevronDown />}
       </S.Preview>
